+++ py_meta_utils.py
@@ -35,6 +35,15 @@ class McsArgs:
     def Meta(self) -> object:
         return self.clsdict['Meta']
 
+    @property
+    def is_abstract(self) -> bool:
+        """
+        Whether or not the class-under-construction was declared as abstract (NOTE: this
+        property is usable even *before* the :class:`MetaOptionsFactory` has run)
+        """
+        meta_value = getattr(self.clsdict.get('Meta'), 'abstract', False)
+        return self.clsdict.get(ABSTRACT_ATTR, meta_value) is True
+
     def __iter__(self):
         return iter([self.mcs, self.name, self.bases, self.clsdict])